requirements.txt
streamlit
groq
python-dotenv
requests
email-validator 